--- playground/mocks/data/idp/idx/authenticator-verification-factor-list.json
@@ -1,195 +1,0 @@
-{
-  "version": "1.0.0",
-  "stateHandle": "02PVkP3FJyDnqUKkkxIZhxbsx7a2S-hC1JxIE6AXzp",
-  "expiresAt": "2021-11-29T19:51:07.000Z",
-  "intent": "LOGIN",
-  "remediation": {
-    "type": "array",
-    "value": [
-      {
-        "rel": [
-          "create-form"
-        ],
-        "name": "select-authenticator-authenticate",
-        "href": "http://localhost:3000/idp/idx/challenge",
-        "method": "POST",
-        "produces": "application/ion+json; okta-version=1.0.0",
-        "value": [
-          {
-            "name": "authenticator",
-            "type": "object",
-            "options": [
-              {
-                "label": "Email",
-                "value": {
-                  "form": {
-                    "value": [
-                      {
-                        "name": "id",
-                        "required": true,
-                        "value": "aut1v94GMPPMvkjbp0g4",
-                        "mutable": false
-                      },
-                      {
-                        "name": "methodType",
-                        "required": false,
-                        "value": "email",
-                        "mutable": false
-                      }
-                    ]
-                  }
-                },
-                "relatesTo": "$.authenticatorEnrollments.value[0]"
-              },
-              {
-                  "label": "Okta Verify",
-                  "value": {
-                      "form": {
-                          "value": [
-                              {
-                                  "name": "id",
-                                  "required": true,
-                                  "value": "aut100dEWdPQ0BZzd0g4",
-                                  "mutable": false
-                              },
-                              {
-                                  "name": "methodType",
-                                  "type": "string",
-                                  "required": false,
-                                  "options": [
-                                      {
-                                          "label": "Enter a code",
-                                          "value": "totp"
-                                      },
-                                      {
-                                          "label": "Get a push notification",
-                                          "value": "push"
-                                      }
-                                  ]
-                              }
-                          ]
-                      }
-                  },
-                  "relatesTo": "$.authenticators.value[1]"
-              }
-            ]
-          },
-          {
-            "name": "stateHandle",
-            "required": true,
-            "value": "02PVkP3FJyDnqUKkkxIZhxbsx7a2S-hC1JxIE6AXzp",
-            "visible": false,
-            "mutable": false
-          }
-        ],
-        "accepts": "application/json; okta-version=1.0.0"
-      }
-    ]
-  },
-  "authenticators": {
-    "type": "array",
-    "value": [
-      {
-        "type": "email",
-        "key": "okta_email",
-        "id": "aut1v94GMPPMvkjbp0g4",
-        "displayName": "Email",
-        "methods": [
-          {
-            "type": "email"
-          }
-        ]
-      },
-      {
-        "type": "app",
-        "key": "okta_verify",
-        "id": "aut100dEWdPQ0BZzd0g4",
-        "displayName": "Okta Verify",
-        "methods": [
-          {
-            "type": "push"
-          },
-          {
-            "type": "totp"
-          }
-        ]
-      }
-    ]
-  },
-  "authenticatorEnrollments": {
-    "type": "array",
-    "value": [
-      {
-        "profile": {
-          "email": "i***1@okta.com"
-        },
-        "type": "email",
-        "key": "okta_email",
-        "id": "eae2ep99TyJZdLB8R0g4",
-        "displayName": "Email",
-        "methods": [
-          {
-            "type": "email"
-          }
-        ]
-      },
-      {
-        "profile": {
-          "deviceName": "iPhone"
-        },
-        "type": "app",
-        "key": "okta_verify",
-        "id": "pfd2q8y7Af7r8rKl90g4",
-        "displayName": "Okta Verify",
-        "methods": [
-          {
-            "type": "push"
-          },
-          {
-            "type": "totp"
-          }
-        ]
-      }
-    ]
-  },
-  "user": {
-      "type": "object",
-      "value": {
-          "id": "00u2ep8CG3BouMb750g4",
-          "identifier": "indira.thirumalaimuruan+testuser1@okta.com",
-          "profile": {
-              "firstName": "Test",
-              "lastName": "user1",
-              "timeZone": "America/Los_Angeles",
-              "locale": "en_US"
-          }
-      }
-  },
-  "cancel": {
-    "rel": [
-      "create-form"
-    ],
-    "name": "cancel",
-    "href": "http://localhost:3000/idp/idx/cancel",
-    "method": "POST",
-    "produces": "application/ion+json; okta-version=1.0.0",
-    "value": [
-      {
-        "name": "stateHandle",
-        "required": true,
-        "value": "02PVkP3FJyDnqUKkkxIZhxbsx7a2S-hC1JxIE6AXzp",
-        "visible": false,
-        "mutable": false
-      }
-    ],
-    "accepts": "application/json; okta-version=1.0.0"
-  },
-  "app": {
-    "type": "object",
-    "value": {
-      "name": "okta_enduser",
-      "label": "Okta Dashboard",
-      "id": "0oazy63LqpLSP2B6r0g3"
-    }
-  }
-}
