--- playground/mocks/data/idp/idx/authenticator-verification-okta-verify-factor-list.json
@@ -1,244 +1,0 @@
-{
-  "version": "1.0.0",
-  "stateHandle": "02f9EONFMnQZtAQHl7Gf4Ye-R4mXc8cOhBUSMc7ubQ",
-  "expiresAt": "2021-11-02T22:10:41.000Z",
-  "intent": "LOGIN",
-  "remediation": {
-    "type": "array",
-    "value": [
-      {
-        "rel": [
-          "create-form"
-        ],
-        "name": "authenticator-verification-data",
-        "href": "http://localhost:3000/idp/idx/challenge",
-        "method": "POST",
-        "produces": "application/ion+json; okta-version=1.0.0",
-        "value": [
-          {
-            "name": "authenticator",
-            "label": "Okta Verify",
-            "form": {
-              "value": [
-                {
-                  "name": "id",
-                  "required": true,
-                  "value": "aut1ejvmCE1iWalJR0g4",
-                  "mutable": false
-                },
-                {
-                  "name": "methodType",
-                  "required": true,
-                  "options": [
-                    {
-                      "value": "push",
-                      "label": "Get a push notification"
-                    },
-                    {
-                      "value": "totp",
-                      "label" :  "Enter a code"
-                    },
-                    {
-                      "value": "signed_nonce",
-                      "label":  "Use Okta FastPass"
-                    }
-                  ]
-                },
-                {
-                  "name" : "autoChallenge",
-                  "type" : "boolean",
-                  "required" : "false",
-                  "value" : true,
-                  "label" : "Send push automatically",
-                  "visible" : true,
-                  "mutable" : true
-                }
-              ]
-            }
-          },
-          {
-            "name": "stateHandle",
-            "required": true,
-            "value": "02f9EONFMnQZtAQHl7Gf4Ye-R4mXc8cOhBUSMc7ubQ",
-            "visible": false,
-            "mutable": false
-          }
-        ],
-        "accepts": "application/json; okta-version=1.0.0"
-      }
-    ]
-  },
-  "currentAuthenticator": {
-    "type": "object",
-    "value": {
-      "resend": {
-          "rel": [
-          "create-form"
-          ],
-          "name": "resend",
-          "href": "http://localhost:3000/idp/idx/challenge",
-          "method": "POST",
-          "produces": "application/ion+json; okta-version=1.0.0",
-          "value": [
-          {
-              "name": "authenticator",
-              "required": true,
-              "value": {
-              "methodType": "push",
-              "id": "aut1ejvmCE1iWalJR0g4"
-              },
-              "visible": false,
-              "mutable": false
-          },
-          {
-              "name": "stateHandle",
-              "required": true,
-              "value": "02f9EONFMnQZtAQHl7Gf4Ye-R4mXc8cOhBUSMc7ubQ",
-              "visible": false,
-              "mutable": false
-          }
-          ],
-          "accepts": "application/json; okta-version=1.0.0"
-      },
-      "type": "app",
-      "key": "okta_verify",
-      "id": "aut1ejvmCE1iWalJR0g4",
-      "displayName": "Okta Verify",
-      "methods": [
-        {
-          "type": "push"
-        },
-        {
-            "type": "totp"
-        },
-        {
-          "type": "signed_nonce"
-        }
-      ]
-    }
-  },
-  "authenticators": {
-    "type": "array",
-    "value": [
-      {
-        "type": "app",
-        "key": "okta_verify",
-        "id": "aut1ejvmCE1iWalJR0g4",
-        "displayName": "Okta Verify",
-        "methods": [
-          {
-            "type": "push"
-          },
-          {
-            "type": "totp"
-          },
-          {
-            "type": "signed_nonce"
-          }
-        ]
-      },
-      {
-        "type": "phone",
-        "key": "phone_number",
-        "id": "aut1ejsBoUqlkTqXq0g4",
-        "displayName": "Phone",
-        "methods": [
-          {
-            "type": "sms"
-          }
-        ]
-      }
-    ]
-  },
-  "authenticatorEnrollments": {
-    "type": "array",
-    "value": [
-      {
-        "profile": {
-          "email": "t***1@test.com"
-        },
-        "type": "email",
-        "key": "okta_email",
-        "id": "eae1ertKRVa9V1yKf0g4",
-        "displayName": "Email",
-        "methods": [
-          {
-            "type": "email"
-          }
-        ]
-      },
-      {
-        "profile": {
-          "deviceName": "Sanjay's Note10+"
-        },
-        "type": "app",
-        "key": "okta_verify",
-        "id": "pfd1f7bKSou3XpIAr0g4",
-        "displayName": "Okta Verify",
-        "methods": [
-          {
-              "type": "push"
-          },
-          {
-              "type": "totp"
-          }
-        ]
-      },
-      {
-        "profile": {
-          "phoneNumber": "+1 XXX-XXX-3657"
-        },
-        "type": "phone",
-        "key": "phone_number",
-        "id": "sms1fz709joi5xHhg0g4",
-        "displayName": "Phone",
-        "methods": [
-          {
-            "type": "sms"
-          }
-        ]
-      }
-    ]
-  },
-  "user": {
-    "type": "object",
-    "value": {
-      "id": "00u1ersTtSrHyIlwP0g4",
-      "identifier": "test1@test.com",
-      "profile": {
-          "firstName": "test",
-          "lastName": "test",
-          "timeZone": "America/Los_Angeles",
-          "locale": "en_US"
-      }
-    }
-  },
-  "cancel": {
-    "rel": [
-      "create-form"
-    ],
-    "name": "cancel",
-    "href": "http://localhost:3000/idp/idx/cancel",
-    "method": "POST",
-    "produces": "application/ion+json; okta-version=1.0.0",
-    "value": [
-      {
-        "name": "stateHandle",
-        "required": true,
-        "value": "02f9EONFMnQZtAQHl7Gf4Ye-R4mXc8cOhBUSMc7ubQ",
-        "visible": false,
-        "mutable": false
-      }
-    ],
-    "accepts": "application/json; okta-version=1.0.0"
-  },
-  "app": {
-    "type": "object",
-    "value": {
-      "name": "okta_enduser",
-      "label": "Okta Dashboard",
-      "id": "0oa1ehulalePYCyiK0g4"
-    }
-  }
-}
-
