--- playground/mocks/data/idp/idx/identify-with-username-password.json
@@ -1,113 +1,0 @@
-{
-  "version": "1.0.0",
-  "stateHandle": "eyJ6aXAiOiJERUYiLCJhbGlhcyI6ImVuY3J5cHRpb25rZXkiLC",
-  "expiresAt": "2021-11-29T21:32:14.000Z",
-  "intent": "LOGIN",
-  "remediation": {
-    "type": "array",
-    "value": [
-      {
-        "rel": [
-          "create-form"
-        ],
-        "name": "identify",
-        "href": "http://localhost:3000/idp/idx/identify",
-        "method": "POST",
-        "produces": "application/ion+json; okta-version=1.0.0",
-        "value": [
-          {
-            "name": "identifier",
-            "label": "Username"
-          },
-          {
-            "name": "credentials",
-            "type": "object",
-            "form": {
-              "value": [
-                {
-                  "name": "passcode",
-                  "label": "Password",
-                  "secret": true
-                }
-              ]
-            },
-            "required": true
-          },
-          {
-            "name": "rememberMe",
-            "type": "boolean",
-            "label": "Remember this device"
-          },
-          {
-            "name": "stateHandle",
-            "required": true,
-            "value": "eyJ6aXAiOiJERUYiLCJhbGlhcyI6ImVuY3J5cHRpb25rZXkiLC",
-            "visible": false,
-            "mutable": false
-          }
-        ],
-        "accepts": "application/json; okta-version=1.0.0"
-      }
-    ]
-  },
-  "currentAuthenticator": {
-    "type": "object",
-    "value": {
-      "recover": {
-        "rel": [
-          "create-form"
-        ],
-        "name": "recover",
-        "href": "http://localhost:3000/idp/idx/recover",
-        "method": "POST",
-        "produces": "application/ion+json; okta-version=1.0.0",
-        "value": [
-          {
-            "name": "stateHandle",
-            "required": true,
-            "value": "eyJ6aXAiOiJERUYiLCJhbGlhcyI6ImVuY3",
-            "visible": false,
-            "mutable": false
-          }
-        ],
-        "accepts": "application/json; okta-version=1.0.0"
-      },
-      "type": "password",
-      "key": "okta_password",
-      "id": "aut1v937PAvvblLV50g4",
-      "displayName": "Password",
-      "methods": [
-        {
-          "type": "password"
-        }
-      ]
-    }
-  },
-  "cancel": {
-    "rel": [
-      "create-form"
-    ],
-    "name": "cancel",
-    "href": "http://localhost:3000/idp/idx/cancel",
-    "method": "POST",
-    "produces": "application/ion+json; okta-version=1.0.0",
-    "value": [
-      {
-        "name": "stateHandle",
-        "required": true,
-        "value": "eyJ6aXAiOiJERUYiLCJhbGlhcyI6ImVuY3",
-        "visible": false,
-        "mutable": false
-      }
-    ],
-    "accepts": "application/json; okta-version=1.0.0"
-  },
-  "app": {
-    "type": "object",
-    "value": {
-      "name": "okta_enduser",
-      "label": "Okta Dashboard",
-      "id": "0oazy63LqpLSP2B6r0g3"
-    }
-  }
-}
