--- src/v2/view-builder/views/ov/ChallengeAuthenticatorOktaVerifyView.js
@@ -1,24 +1,0 @@
-import { BaseFooter } from '../../internals';
-import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
-import { 
-  Body as ChallengeAuthenticatorDataOktaVerifyPushOnlyView
-} from './ChallengeAuthenticatorDataOktaVerifyPushOnlyView';
-import { Body as ChallengeAuthenticatorDataOktaVerifyView } from './ChallengeAuthenticatorDataOktaVerifyView';
-
-export default BaseAuthenticatorView.extend({
-  initialize() {
-    BaseAuthenticatorView.prototype.initialize.apply(this, arguments);
-    if (this.isPushOnlyFlow()) {
-      this.Body = ChallengeAuthenticatorDataOktaVerifyPushOnlyView;
-    } else {
-      this.Body = ChallengeAuthenticatorDataOktaVerifyView;
-      this.Footer = BaseFooter;
-    }
-  },
-
-  isPushOnlyFlow() {
-    const methodType = this.options.appState.getSchemaByName('authenticator.methodType');
-    const methodTypeOptions = methodType?.options;
-    return methodTypeOptions.length === 1 && methodTypeOptions[0].value === 'push';
-  },
-});
