+++ test/testcafe/framework/page-objects/BasePageObject.js
@@ -7,6 +7,7 @@ const SKIP_LINK = '.auth-footer .js-skip';
 const SKIP_SET_UP_LINK = '.auth-footer .js-skip-setup';
 const SWITCH_AUTHENTICATOR_LINK = '.auth-footer .js-switchAuthenticator';
 const ionMessagesSelector = '.ion-messages-container';
+const SUBTITLE_SELECTOR = '[data-se="o-form-explain"]';
 
 export default class BasePageObject {
   constructor(t) {
@@ -161,5 +162,9 @@ export default class BasePageObject {
 
   refresh() {
     return this.t.eval(() => location.reload(true));
+  }
+
+  subtitleExists() {
+    return this.form.elementExist(SUBTITLE_SELECTOR);
   }
 }
